pyproject.toml
[project]
name = "indexv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.1",
    "llama-index>=0.14.12",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-storage-docstore-redis>=0.4.1",
    "llama-index-storage-kvstore-redis>=0.4.1",
    "llama-index-vector-stores-chroma>=0.5.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
[tool.pytest.ini_options]
pythonpath = "."
